-Stock-Trading-Agents-with-CrewAI-main/AI_Powered_Stock _Trading_Agent/requirements.txt
crewai==0.140.0
crewai-tools==0.49.0
python-dotenv==1.1.1
yfinance==0.2.64
pandas>=1.5.0
numpy>=1.21.0
feedparser>=6.0.0
requests>=2.28.0
tweepy>=4.14.0
